rpc: int64 is just int

+++ src/lib/rpc.ml
@@ -171,7 +171,7 @@ let rec to_string t =
 
 let rpc_of_t x = x
 
-let rpc_of_int64 i = Int64 i
+let rpc_of_int64 i = Int i
 
 let rpc_of_int32 i = Int32 i
